Construct webpack config in the config file
It's counterintiutive to go to webpack.config.js and not find the
top-level config properties that we expect. Moving everything into the
webpack helpers util was an overeager "optimization".

Move webpack helper invocations back into the webpack config, but keep
their definitions in the webpack helpers util. Remove the helper that
constructed the base config.